requirements.txt
bibtexparser==1.4.3
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pyparsing==3.2.3
pyperclip==1.9.0
requests==2.32.4
urllib3==2.5.0
citeproc-py>=0.6.0
